Split out protocol to prepare for interface segregation
